+++ fabric/src/main/java/com/slymask3/instantblocks/InstantBlocks.java
@@ -17,12 +17,16 @@ import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
+import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.core.BlockPos;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 
 public class InstantBlocks implements ModInitializer {
     @Override
@@ -64,6 +68,11 @@ public class InstantBlocks implements ModInitializer {
         public void sendToClient(Player player, AbstractPacket message) {
             if(Helper.isServer(player.getLevel())) {
                 ServerPlayNetworking.send((ServerPlayer)player, message.getKey(), message.getBuffer());
+            }
+        }
+        public void sendToAllAround(Level world, BlockPos pos, AbstractPacket message) {
+            for(ServerPlayer player : PlayerLookup.tracking((ServerLevel)world, pos)) {
+                ServerPlayNetworking.send(player, message.getKey(), message.getBuffer());
             }
         }
     }